Add files via upload
We have built a pedestrian detection system for a self-driving car. The car-pedestrian detection system should identify exactly where the people are walking so that we can take care of them or keep them away. So the car system will create a bounding box around these people so that the car can pinpoint where in the image people are, and then accordingly decide as to which path to take, to avoid any mishaps. Instead of taking a patch from the original image, we are passing the original image through a neural network to reduce dimensions. We have used ImageAI, a python library that supports the most recent machine learning algorithms for computer vision tasks. in this project, we have downloaded 'RetinaNet' pretrained model.